move unifiedruleresults interfaces to another file
