+++ src/common/components/cards/failed-instances-section.tsx
@@ -3,8 +3,7 @@
 import { NamedFC } from 'common/react/named-fc';
 import * as React from 'react';
 
-import { TargetAppData, UnifiedResult, UnifiedRuleResultStatus } from '../../../common/types/store-data/unified-data-interface';
-import { GuidanceLink } from '../../../scanner/rule-to-links-mappings';
+import { TargetAppData, UnifiedStatusResults } from '../../../common/types/store-data/unified-data-interface';
 import { UserConfigurationStoreData } from '../../types/store-data/user-configuration-store';
 import { ResultSection, ResultSectionDeps } from './result-section';
 
@@ -14,18 +13,6 @@ export type FailedInstancesSectionProps = {
     ruleResultsByStatus: UnifiedStatusResults;
     userConfigurationStoreData: UserConfigurationStoreData;
     targetAppInfo: TargetAppData;
-};
-
-export interface UnifiedRuleResult {
-    id: string;
-    nodes: UnifiedResult[];
-    description: string;
-    url: string;
-    guidance: GuidanceLink[];
-}
-
-export type UnifiedStatusResults = {
-    [key in UnifiedRuleResultStatus]: UnifiedRuleResult[];
 };
 
 export const FailedInstancesSection = NamedFC<FailedInstancesSectionProps>(
