+++ src/common/types/store-data/unified-data-interface.ts
@@ -86,4 +86,16 @@ export type InstanceResultStatus = 'pass' | 'fail' | 'unknown';
 
 export type UnifiedRuleResultStatus = InstanceResultStatus | 'inapplicable';
 
+export interface UnifiedRuleResult {
+    id: string;
+    nodes: UnifiedResult[];
+    description: string;
+    url: string;
+    guidance: GuidanceLink[];
+}
+
+export type UnifiedStatusResults = {
+    [key in UnifiedRuleResultStatus]: UnifiedRuleResult[];
+};
+
 export const AllRuleResultStatuses: UnifiedRuleResultStatus[] = ['pass', 'fail', 'unknown', 'inapplicable'];
